+++ src/test/java/nom/bdezonia/zorbage/type/data/float16/quaternion/TestQuaternionFloat16Algebra.java
@@ -195,7 +195,7 @@ public class TestQuaternionFloat16Algebra {
 		assertEquals(6, b.k(), 0);
 		
 		// G.OHLF.cos();
-		a.setR(Math.PI/2);
+		a.setR((float)Math.PI/2);
 		a.setI(0);
 		a.setJ(0);
 		a.setK(0);
@@ -206,7 +206,7 @@ public class TestQuaternionFloat16Algebra {
 		assertEquals(0, b.k(), 0);
 		
 		// G.OHLF.cosh();
-		a.setR(Math.PI/2);
+		a.setR((float)Math.PI/2);
 		a.setI(0);
 		a.setJ(0);
 		a.setK(0);
@@ -221,7 +221,7 @@ public class TestQuaternionFloat16Algebra {
 		a.setI(2);
 		a.setJ(3);
 		a.setK(4);
-		b.setR(0.5);
+		b.setR(0.5f);
 		b.setI(0);
 		b.setJ(0);
 		b.setK(0);
@@ -392,10 +392,10 @@ public class TestQuaternionFloat16Algebra {
 		assertEquals(1, t.v(), 0);
 		
 		// G.OHLF.round();
-		a.setR(1.1);
-		a.setI(-2.2);
-		a.setJ(3.3);
-		a.setK(-4.4);
+		a.setR(1.1f);
+		a.setI(-2.2f);
+		a.setJ(3.3f);
+		a.setK(-4.4f);
 		t.setV(1);
 		G.QHLF.round().call(Mode.HALF_EVEN, t, a, b);
 		assertEquals(1, b.r(), 0);
@@ -404,10 +404,10 @@ public class TestQuaternionFloat16Algebra {
 		assertEquals(-4, b.k(), 0);
 		
 		// G.OHLF.scale();
-		a.setR(1.1);
-		a.setI(-2.2);
-		a.setJ(3.3);
-		a.setK(-4.4);
+		a.setR(1.1f);
+		a.setI(-2.2f);
+		a.setJ(3.3f);
+		a.setK(-4.4f);
 		b.setR(3);
 		b.setI(0);
 		b.setJ(0);
@@ -419,7 +419,7 @@ public class TestQuaternionFloat16Algebra {
 		assertEquals(-13.2, c.k(), tol);
 		
 		// G.OHLF.sin();
-		a.setR(Math.PI/2);
+		a.setR((float)Math.PI/2);
 		a.setI(0);
 		a.setJ(0);
 		a.setK(0);
@@ -440,7 +440,7 @@ public class TestQuaternionFloat16Algebra {
 		// G.OHLF.sincpi();
 		
 		// G.OHLF.sinh();
-		a.setR(Math.PI/2);
+		a.setR((float)Math.PI/2);
 		a.setI(0);
 		a.setJ(0);
 		a.setK(0);
@@ -479,7 +479,7 @@ public class TestQuaternionFloat16Algebra {
 		assertEquals(-7, c.k(), 0);
 		
 		// G.OHLF.tan();
-		a.setR(Math.PI/4);
+		a.setR((float)Math.PI/4);
 		a.setI(0);
 		a.setJ(0);
 		a.setK(0);
@@ -490,7 +490,7 @@ public class TestQuaternionFloat16Algebra {
 		assertEquals(0, b.k(), 0);
 		
 		// G.OHLF.tanh();
-		a.setR(Math.PI/4);
+		a.setR((float)Math.PI/4);
 		a.setI(0);
 		a.setJ(0);
 		a.setK(0);
